ndctl/namespace: Clarify that 'reconfigure' == 'destroy+create'
The 'reconfigure' option to create-namespace is simply a shortcut for
the following flow:

- Read all parameters from @victim_namespace
- Destroy @victim_namespace
- Create @new_namespace merging old parameters with new ones

Critically this process makes no attempt to preserve data. In fact
previous data is almost always destroyed except for cases where the mode
remains 'raw' over the transition. Even there the create step may choose
new free space for @new_namespace that was not previously allocated to
@victim_namespace. For reconfigurations where the target mode is 'fsdax'
or 'devdax' any size increase may increase the size of metadata needed
at the start of the namespace and overwrite previous data from
@victim_namespace.

Highlight these details in the man page to preclude assumptions that the
a namespace can be reconfigured while keeping existing data intact.

Reported-by: Seema Pandit <seema.pandit@intel.com>
Reported-by: Bharath Venkatesh <bharath.venkatesh@intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>