ndctl/lib: fix symbol redefinitions reported by GCC10
A toolchain update in Fedora 32 caused new compile errors due to
multiple definitions of dimm_ops structures. The declarations in
'private.h' for the various NFIT families are present so that libndctl
can find all the per-family dimm-ops. However they need to be declared
as extern because the actual definitions are in <family>.c.
Additionally, 'param' instances in list.c and monitor.c need to be
marked as static.

Cc: Dan Williams <dan.j.williams@intel.com>
Cc: Eric Sandeen <sandeen@redhat.com>
Suggested-by: Dan Williams <dan.j.williams@intel.com>
Reviewed-by: Dan Williams <dan.j.williams@intel.com>
Acked-by: Jeff Moyer <jmoyer@redhat.com>
Link: https://lore.kernel.org/r/20200131200630.17582-1-vishal.l.verma@intel.com
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>