ndctl/namespace: Validate resource alignment for dax-mode namespaces
The kernel sets the default region alignment to 16M to promote
cross-arch compatible namespace creation. While ndctl never touches the
region alignment the end user might have changed it from its default.
Enforce 16MiB alignment for the namespace resource base by default for
dax-mode namespaces.

It is still possible to use a 2MiB region-align for dax-mode namespaces
on x86, but that requires --force to bypass this default alignment
check.

I chose a hard coded default value in ndctl with a --force to bypass the
check rather than having a new sysfs attribute to probe for this detail.
I.e. the kernel could export the minimum alignment for dax namespaces,
but since the minimum compat value is already known, no need for a trip
to the kernel.

Reviewed-by: Vishal Verma <vishal.l.verma@intel.com>
Cc: Aneesh Kumar K.V <aneesh.kumar@linux.ibm.com>
Cc: Jeff Moyer <jmoyer@redhat.com>
Link: https://lore.kernel.org/r/158300766271.2141307.12541085220843554138.stgit@dwillia2-desk3.amr.corp.intel.com
Signed-off-by: Dan Williams <dan.j.williams@intel.com>