ndctl/list: Add option to list configured + disabled namespaces
When inspecting namespaces multiple tools require the namespace to be
disabled so that the tool can change modes to inspect namespace
metadata, for example read-infoblock and check-namespace.

While a namespace is disabled it can be listed with "--idle", but that
will also include seed namespaces in the output. Add a --configured
option that lists includes namespaces with non-zero size in the listing.
For regions, dimms, and buses, it is equivalent to --idle.

Reviewed-by: Vishal Verma <vishal.l.verma@intel.com>
Link: https://lore.kernel.org/r/158300778534.2141307.3858623337643122737.stgit@dwillia2-desk3.amr.corp.intel.com
Signed-off-by: Dan Williams <dan.j.williams@intel.com>