ndctl/bus: Add 'activate-firmware' command
For platforms where firmware can be activated at runtime, trigger
that activation. The activation method, or system context while the
activation is performed is determined by the driver between "suspend" and
"live". Where "live" has no memory controller side-effects during the
activation and "suspend" indicates that the platform will take the memory
controller offline for a short period of time.

An override, to attempt "live" activation, is implemented in the --force
option to activate-firmware.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>