test/daxctl-create.sh: Validate @size versus mappingX sizes
Check that the size calculated by mappingNNNN is the same as advertised
by the sysfs size attribute, thus looping through all the mappings.

Perhaps it would be enough to have just such validation in setup_dev()
to catch the bug in [0]. Nevertheless, validate the test cases where a
certain number of mappings are meant to be created.

This may be too tightly coupled to how the kernel allocates space from a
region. If this logic changes in the future, this test may need to be
relaxed or altered accordingly.

[0]: https://lore.kernel.org/linux-nvdimm/CAPcyv4hFS7JS9s7cUY=2Ru2kUTRsesxwX1PGnnc_tudJjoDUGw@mail.gmail.com/

Signed-off-by: Joao Martins <joao.m.martins@oracle.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>
Link: https://lore.kernel.org/linux-nvdimm/5cf76b3d-21db-daf9-dc1d-d38714a9a7c2@oracle.com/