libndctl: Fix probe of non-nfit nvdimms
commit 107a24ff429f ("ndctl/list: Add firmware activation
enumeration") introduced changes in add_dimm() to enumerate the status
of firmware activation. However it broke the probe for non-nfit nvdimms
like one provided by papr-scm. This cause an errors like:

  # ndctl list
  libndctl: add_dimm: nmem0: probe failed: No such device
  libndctl: __sysfs_device_parse: nmem0: add_dev() failed

Fix this by removing the offending branch in add_dimm(). This continues
the flow of add_dimm() probe even if the nfit is not detected on the
associated bus.

Fixes: 107a24ff429fa("ndctl/list: Add firmware activation enumeration")
Signed-off-by: Vaibhav Jain <vaibhav@linux.ibm.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>
Link: https://lore.kernel.org/r/20201009120009.243108-1-vaibhav@linux.ibm.com