ndctl/namespace: Reconfigure in-place
While it has been documented that the namespace reconfiguration process
involves destroying and recreating a namespace, the kernel device changing
as result of a reconfigure operation is surprising. For example a sequence
like:

ndctl create-namespace -s 4G
ndctl create-namespace -s 4G
ndctl create-namespace -s 4G

Results in 3 namespaces:

namespace0.0
namespace0.1
namespace0.2

...whereby after each creation step the next seed is incremented. At the
end of this process namespace0.3 is the next namespace that will be
created. However, when reconfiguration also picks the seed for the target
vessel of the new namespace configuration it is unexpected that:

ndctl create-namespace -e namespace0.1 -m sector

...results in namespace0.3 becoming enabled.

Teach the reconfigure path in ndctl to try to reuse the existing kernel
device whenever possible.

Reported-by: Steve Scargall <steve.scargall@intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>
Link: https://github.com/pmem/ndctl/issues/152
Link: https://lore.kernel.org/r/160583984571.3214303.18399638980700230679.stgit@dwillia2-desk3.amr.corp.intel.com