ndctl, util: add parse-configs helper
Add parse-config util to help ndctl commands parse ndctl global
configuration files. This provides a parse_configs_prefix() helper which
uses the iniparser APIs to read all applicable config files, and either
return a 'value' for a requested 'key', or perform a callback if
requested. The operation is defined by a 'struct config' which
encapsulates the key to search for, the location to store the value, and
any callbacks to be executed.

Link: https://lore.kernel.org/r/20211210223440.3946603-2-vishal.l.verma@intel.com
Reviewed-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: QI Fuli <qi.fuli@fujitsu.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>