ndctl/test: make inject-smart.sh more tolerant of decimal fields
Some combinations of json-c/jq/other libraries seem to produce differing
outputs for the final jq-filtered smart fields, in that some have a
decimal "42.0" numeric field, where as in other combinations it is a
simple "42" (still a numeric field, not string).

This shouldn't matter in practice, but for this contrived test case, we
need to make sure that "42" is treated the same as "42.0"

Normalize all fields before comparing them to "%0.0f" so that the
comparison doesn't result in superfluous failures.

Reported-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>