monitor: Fix the monitor config file parsing
Presently, ndctl monitor is not parsing both the default and user specified
config files. The behaviour is quitely masked with the recent iniparser
parsing code without any signs until you remove the /etc/ndctl.conf.d
directory when the command fails as,

 #ndctl monitor -d nmem0
 iniparser: cannot open /etc/ndctl.conf.d

The error is coming from the libiniparser when the monitor parser is
initialised with MONITOR_CALLBACK type with parse_monitor_config() as the
callback. The configs->key is set to the NDCTL_CONF_FILE for this.
The parse_config_file() compares the filename with the key before
calling the custom callback. The current code calls the
parse_config_prefix() with either the default directory path or the
custom config filepath(i.e -c <XYZ>) while the configs->key is set to
the NDCTL_CONF_FILE. Since both these strings don't match the
NDCTL_CONF_FILE, parse_monitor_config() is not called at all and instead
generic iniparser code path is taken.

The patch sets the config key to the correct filename before the calls to
parse_config_prefix().

The previous behaviour for missing monitor.conf file in the default path
was to ignore and continue. The current callback parse_monitor_config()
reports an error as e889fa5e removed the chunk taking care of this case.

So the patch gets the "current" config directory and checks if the default
monitor.conf file exists, dont attempt to parse if the file is missing.

Link: https://lore.kernel.org/r/164750955519.2000193.16903542741359443926.stgit@LAPTOP-TBQTPII8
Fixes: e889fa5ef7ff ("ndctl, monitor: refator monitor for supporting multiple config files")
Signed-off-by: Shivaprasad G Bhat <sbhat@linux.ibm.com>
Signed-off-by: Vaibhav Jain <vaibhav@linux.ibm.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>