scripts: update release helper scripts for meson and cxl
The prepare-release.sh and do_abidiff scripts perform sanity checking
for library versioning and also guard against accidental ABI breakage
by comparing the current release with the previous using 'abipkgdiff'
from libabigail. Teach the scripts about libcxl, so that it too can
participate in the above checks.

Additionally, move the checks over to the new meson regime. This does
break any checking for the older autotools based build, but that should
be okay.

Link: https://lore.kernel.org/r/20220615224813.523053-6-vishal.l.verma@intel.com
Cc: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>